submits/copy

See merge request 2023-se/se-226-2-Nikita-Oleynik-oleynik_n_a!47

+++ tasks/copy/test.cpp
@@ -5,7 +5,7 @@
 
 template <class T>
 void Equal(const std::vector<T>& actual, const std::vector<T>& expected) {
-    for (int i = 0; i < actual.size(); ++i) {
+    for (size_t i = 0; i < actual.size(); ++i) {
         REQUIRE(actual[i] == expected[i]);
     }
 }
